requirements.txt
claude-agent-sdk
pyperclip
python-dotenv
supabase
atlassian-python-api
streamlit